requirements.txt
pydantic==1.10.5
python-telegram-bot==20.1
redis==4.5.1
